+++ demisto_sdk/commands/common/tools.py
@@ -1086,3 +1086,17 @@ def get_content_file_type_dump(file_path: str) -> Callable[[str], str]:
     elif file_extension == '.json':
         curr_string_transformer = partial(json.dumps, indent=4)
     return curr_string_transformer
+
+
+def get_code_lang(file_data: dict, file_entity: str) -> str:
+    """
+    Returns the code language by the file entity
+    :param file_data: The file data
+    :param file_entity: The file entity
+    :return: The code language
+    """
+    if file_entity == INTEGRATIONS_DIR:
+        return file_data.get('script', {}).get('type', '')
+    elif file_entity == SCRIPTS_DIR:
+        return file_data.get('type', {})
+    return ''